+++ src/main/java/com/github/derickfelix/Constants.java
@@ -25,6 +25,16 @@ class Constants {
     }
   };
 
+  public static enum JOIN {
+    INNER(" INNER JOIN "), LEFT(" LEFT JOIN "), RIGHT(" RIGHT JOIN ");
+
+    public String joinOpt;
+
+    private JOIN(String joinOpt) {
+      this.joinOpt = joinOpt;
+    }
+  }
+
   Constants() {
     dialectConstants.put(SqlDialect.Sql.sqlDialect + Constants.Actions.SEPARATOR.action, "");
     dialectConstants.put(Constants.SqlDialect.Postgres.sqlDialect + Constants.Actions.SEPARATOR.action, "`");
